Moving BasicGrid and ColorCell to main Cells file; adjusting imports
